Moved fight functionality from ShipService to new FightService.
