+++ service/src/main/java/restwars/service/ServiceModule.java
@@ -6,6 +6,8 @@ import restwars.service.building.BuildingService;
 import restwars.service.building.impl.BuildingServiceImpl;
 import restwars.service.event.EventService;
 import restwars.service.event.impl.EventServiceImpl;
+import restwars.service.fight.FightService;
+import restwars.service.fight.impl.FightServiceImpl;
 import restwars.service.flight.FlightService;
 import restwars.service.flight.impl.FlightServiceImpl;
 import restwars.service.infrastructure.RandomNumberGenerator;
@@ -93,6 +95,11 @@ public class ServiceModule {
     @Provides
     FlightService providesFlightService(FlightServiceImpl flightService) {
         return flightService;
+    }
+
+    @Provides
+    FightService providesFightService(FightServiceImpl fightService) {
+        return fightService;
     }
 
     @Provides
