+++ service/src/main/java/restwars/service/fight/Fight.java
@@ -1,8 +1,9 @@
-package restwars.service.ship;
+package restwars.service.fight;
 
 import com.google.common.base.Objects;
 import com.google.common.base.Preconditions;
 import restwars.service.resource.Resources;
+import restwars.service.ship.Ships;
 
 import java.util.UUID;
