+++ service/src/main/java/restwars/service/fight/FightWithPlanetAndPlayer.java
@@ -1,4 +1,4 @@
-package restwars.service.ship;
+package restwars.service.fight;
 
 import com.google.common.base.Objects;
 import com.google.common.base.Preconditions;
