+++ service/src/main/java/restwars/service/flight/DetectedFlight.java
@@ -1,4 +1,4 @@
-package restwars.service.ship;
+package restwars.service.flight;
 
 import com.google.common.base.Objects;
 import com.google.common.base.Preconditions;
