+++ service/src/main/java/restwars/service/flight/DetectedFlightWithSender.java
@@ -3,7 +3,6 @@ package restwars.service.flight;
 import com.google.common.base.Objects;
 import com.google.common.base.Preconditions;
 import restwars.service.player.Player;
-import restwars.service.ship.DetectedFlight;
 
 public class DetectedFlightWithSender {
     private final DetectedFlight detectedFlight;
