+++ service/src/main/java/restwars/service/flight/impl/flighthandler/AttackFlightHandler.java
@@ -5,6 +5,8 @@ import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 import restwars.mechanics.ShipMechanics;
 import restwars.service.event.EventService;
+import restwars.service.fight.Fight;
+import restwars.service.fight.FightDAO;
 import restwars.service.flight.DetectedFlightDAO;
 import restwars.service.flight.Flight;
 import restwars.service.flight.FlightDAO;
@@ -14,7 +16,9 @@ import restwars.service.infrastructure.UUIDFactory;
 import restwars.service.planet.Planet;
 import restwars.service.planet.PlanetDAO;
 import restwars.service.resource.Resources;
-import restwars.service.ship.*;
+import restwars.service.ship.Hangar;
+import restwars.service.ship.HangarDAO;
+import restwars.service.ship.Ships;
 
 import java.util.Optional;
 import java.util.UUID;
