+++ service/src/main/java/restwars/service/flight/impl/flighthandler/FightCalculator.java
@@ -4,10 +4,10 @@ import com.google.common.base.Preconditions;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 import restwars.mechanics.ShipMechanics;
+import restwars.service.fight.Fight;
 import restwars.service.infrastructure.RandomNumberGenerator;
 import restwars.service.infrastructure.UUIDFactory;
 import restwars.service.resource.Resources;
-import restwars.service.ship.Fight;
 import restwars.service.ship.Ship;
 import restwars.service.ship.ShipType;
 import restwars.service.ship.Ships;
