+++ service/src/main/java/restwars/service/ship/ShipService.java
@@ -4,8 +4,6 @@ import restwars.service.planet.Planet;
 import restwars.service.player.Player;
 
 import java.util.List;
-import java.util.Optional;
-import java.util.UUID;
 
 public interface ShipService {
     /**
@@ -40,15 +38,6 @@ public interface ShipService {
     Ships findShipsOnPlanet(Planet planet);
 
     /**
-     * Finds all fights where the given player is either the attacker or the defender which happend since the given round.
-     *
-     * @param player Player.
-     * @param round  Round (inclusive).
-     * @return Fights.
-     */
-    List<FightWithPlanetAndPlayer> findFightsWithPlayerSinceRound(Player player, long round);
-
-    /**
      * Manifests the given ships on the given planet.
      *
      * @param player Player.
@@ -56,12 +45,4 @@ public interface ShipService {
      * @param ships  Ships.
      */
     void manifestShips(Player player, Planet planet, Ships ships);
-
-    /**
-     * Finds the fight with the given id.
-     *
-     * @param id Id.
-     * @return Fight, if found.
-     */
-    Optional<FightWithPlanetAndPlayer> findFight(UUID id);
 }
