+++ storage/src/main/java/restwars/storage/JooqDAOModule.java
@@ -5,12 +5,12 @@ import dagger.Provides;
 import restwars.service.building.BuildingDAO;
 import restwars.service.building.ConstructionSiteDAO;
 import restwars.service.event.EventDAO;
+import restwars.service.fight.FightDAO;
 import restwars.service.flight.DetectedFlightDAO;
 import restwars.service.flight.FlightDAO;
 import restwars.service.infrastructure.RoundDAO;
 import restwars.service.planet.PlanetDAO;
 import restwars.service.player.PlayerDAO;
-import restwars.service.ship.FightDAO;
 import restwars.service.ship.HangarDAO;
 import restwars.service.ship.ShipInConstructionDAO;
 import restwars.service.technology.ResearchDAO;
