+++ storage/src/main/java/restwars/storage/fight/JooqFightDAO.java
@@ -7,9 +7,14 @@ import org.jooq.Result;
 import org.jooq.SelectOnConditionStep;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
+import restwars.service.fight.Fight;
+import restwars.service.fight.FightDAO;
+import restwars.service.fight.FightWithPlanetAndPlayer;
 import restwars.service.planet.Planet;
 import restwars.service.resource.Resources;
-import restwars.service.ship.*;
+import restwars.service.ship.Ship;
+import restwars.service.ship.ShipType;
+import restwars.service.ship.Ships;
 import restwars.service.unitofwork.UnitOfWorkService;
 import restwars.storage.jooq.AbstractJooqDAO;
 import restwars.storage.jooq.tables.Player;
